+++ src/main/java/com/excilys/formation/java/computerdb/dto/CompanyDto.java
@@ -7,22 +7,22 @@ package com.excilys.formation.java.computerdb.dto;
  *
  */
 public class CompanyDto {
-  private int id;
+  private long id;
   private String name;
 
   public CompanyDto() {
   }
 
-  public CompanyDto(int id, String name) {
+  public CompanyDto(long id, String name) {
     this.id = id;
     this.name = name;
   }
 
-  public int getId() {
+  public long getId() {
     return id;
   }
 
-  public void setId(int id) {
+  public void setId(long id) {
     this.id = id;
   }
 
@@ -43,7 +43,7 @@ public class CompanyDto {
   public int hashCode() {
     final int prime = 31;
     int result = 1;
-    result = prime * result + id;
+    result = prime * result + (int)id;
     result = prime * result + ((name == null) ? 0 : name.hashCode());
     return result;
   }
